requirements.txt
# -----------------------------------------------
# Required application dependencies             |
# Run pip install -r requirements.txt           |
# -----------------------------------------------

Flask==2.3.3
flask_cors==3.0.10
flasgger==0.9.7.1
graphene-sqlalchemy==2.3.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
SQLAlchemy==1.4.49
mysqlclient==2.1.1
typing_extensions==4.5.0
flask_graphql==2.0.1
graphql-server-core==1.2.0
python-dotenv==1.0.0
bcrypt==3.2.0
redis==4.6.0
gunicorn
openai
